+++ src/rhodes/__init__.py
@@ -1,4 +1,4 @@
 """Tools to create AWS Step Functions state machines."""
-__version__ = "0.4.0"
+__version__ = "0.5.0"
 
 from .states import StateMachine